apply s390-8214206_fix

+++ hotspot/src/share/vm/utilities/numberSeq.cpp
@@ -369,7 +369,7 @@ BinaryMagnitudeSeq::~BinaryMagnitudeSeq() {
 void BinaryMagnitudeSeq::add(size_t val) {
   Atomic::add(val, &_sum);
 
-  int mag = log2_intptr(val) + 1;
+  int mag = log2_intptr((uintptr_t) val) + 1;
 
   // Defensively saturate for product bits:
   if (mag < 0) {
